Add files via upload
This model states relationships between different tables, how they are connected with each other, with suitable entities in a table. The concept of primary key and foreign key is used as per to create a relationship. If opened in a data modeler environment, it can also generate MySQL syntax to create the database along with tables and relations.